Pulling the updates

+++ mysite/missions/git/view_helpers.py
@@ -16,6 +16,8 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+from django.conf import settings
+
 import mysite.base.unicode_sanity
 from mysite.missions.base.view_helpers import *
 
@@ -29,6 +31,13 @@ class GitRepository(object):
         self.public_url = settings.GIT_REPO_URL_PREFIX + username
 
     def reset(self):
+        if settings.REMOTE_REPO_SETUP_ACCESS_SPEC:
+            subprocess.check_call([
+                'ssh', settings.REMOTE_REPO_SETUP_ACCESS_SPEC,
+                'milestone-a/manage.py', 'git_reset', self.username
+            ])
+            return
+
         if os.path.isdir(self.repo_path):
             shutil.rmtree(self.repo_path)
         subprocess.check_call(['git', 'init', self.repo_path])
@@ -38,7 +47,7 @@ class GitRepository(object):
             ['cp', '../../../missions/git/data/hello.py', '.'], cwd=self.repo_path)
         subprocess.check_call(['git', 'add', '.'], cwd=self.repo_path)
         subprocess.check_call(
-            ['git', 'commit', '-m', '"Initial commit"'], cwd=self.repo_path)
+            ['git', 'commit', '-m', 'Initial commit'], cwd=self.repo_path)
 
         # Touch the git-daemon-export-ok file
         file_obj = file(
@@ -48,6 +57,12 @@ class GitRepository(object):
         person = Person.objects.get(user__username=self.username)
 
     def exists(self):
+        if settings.REMOTE_REPO_SETUP_ACCESS_SPEC:
+            return subprocess.call([
+                'ssh', settings.REMOTE_REPO_SETUP_ACCESS_SPEC,
+                'milestone-a/manage.py', 'git_exists', self.username
+            ]) == 0
+
         return os.path.isdir(self.repo_path)
 
 
@@ -55,6 +70,14 @@ class GitDiffMission(object):
 
     @classmethod
     def commit_if_ok(cls, username, diff):
+        if settings.REMOTE_REPO_SETUP_ACCESS_SPEC:
+            remote_commit = subprocess.Popen([
+                'ssh', settings.REMOTE_REPO_SETUP_ACCESS_SPEC,
+                'milestone-a/manage.py', 'git_commit_if_ok', username
+            ], stdin=subprocess.PIPE)
+            remote_commit.communicate(diff)
+            return remote_commit.returncode == 0
+
         repo = GitRepository(username)
         commit_diff = subprocess.Popen(
             ['git', 'am'], cwd=repo.repo_path, stdin=subprocess.PIPE)
@@ -67,6 +90,6 @@ class GitDiffMission(object):
             subprocess.Popen(
                 ['git', 'commit', '--allow-empty', '-m', commit_msg], cwd=repo.repo_path)
             return True
-        else:
-            subprocess.check_call(['git', 'am', '--abort'], cwd=repo.repo_path)
-            return False
+
+        subprocess.check_call(['git', 'am', '--abort'], cwd=repo.repo_path)
+        return False
